+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.6] - 12/22/2025
+
+### Changes
+
+- Optimization on tab switching
+- Added new tags on project type
+- Fixed: Layout bug on project tab on small screen size
+
 ## [0.1.5] - 12/12/2025
 
 ### Changes